# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Alpogo API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/item/{slug}/{slug_item_type}": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "A\u00f1ade un item al evento",
                "description": "",
                "operationId": "add_item_to_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de un evento",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug_item_type",
                        "in": "path",
                        "description": "Slug de un tipo de item",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "410": {
                        "description": "Item type not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/item/{id}": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Detalle de un item",
                "description": "",
                "operationId": "get_item",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "411": {
                        "description": "Item not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "item"
                ],
                "summary": "Edita un item",
                "description": "",
                "operationId": "update_item",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "411": {
                        "description": "Item not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "item"
                ],
                "summary": "Elimina un item",
                "description": "",
                "operationId": "destroy_item",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "411": {
                        "description": "Item not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "item"
                ],
                "summary": "Edita parcialmente un item",
                "description": "",
                "operationId": "partial_update_item",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    }
                ],
                "responses": {
                    "411": {
                        "description": "Item not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Item": {
            "required": [
                "name",
                "detail",
                "picture",
                "stock",
                "amount",
                "producer_amount"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "El item de Jhon"
                },
                "detail": {
                    "type": "string",
                    "example": "Detalle del item"
                },
                "picture": {
                    "type": "string",
                    "example": "/var/www/html/picture.jpg"
                },
                "stock": {
                    "type": "string",
                    "example": "123"
                },
                "amount": {
                    "type": "string",
                    "example": "123"
                },
                "producer_amount": {
                    "type": "string",
                    "example": "123"
                }
            },
            "type": "object",
            "xml": {
                "name": "Item"
            }
        },
...
    }
}